GRIDEDIT-670 Replace all size_t with UInt

+++ libs/MeshKernel/src/Mesh2D.cpp
@@ -2038,9 +2038,9 @@ meshkernel::Mesh2D Mesh2D::Merge(const Mesh2D& mesh1, const Mesh2D& mesh2)
     // Initialise with mesh1,
     Mesh2D mergedMesh(mesh1);
 
-    size_t mesh1NodeOffset = mesh1.m_nodes.size();
-    size_t mesh1EdgeOffset = mesh1.m_edges.size();
-    size_t mesh1FaceOffset = mesh1.m_numFacesNodes.size();
+    UInt mesh1NodeOffset = static_cast<UInt>(mesh1.m_nodes.size());
+    UInt mesh1EdgeOffset = static_cast<UInt>(mesh1.m_edges.size());
+    UInt mesh1FaceOffset = static_cast<UInt>(mesh1.m_numFacesNodes.size());
 
     // Merge node arrays
     mergedMesh.m_nodes.insert(mergedMesh.m_nodes.end(), mesh2.m_nodes.begin(), mesh2.m_nodes.end());
